services/task-capsule-repo/requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.29
pydantic-settings~=2.2
sqlalchemy[asyncio]~=2.0
asyncpg~=0.29
pyyaml~=6.0
